move modisco funcs into own module
