+++ src/crested/tl/modisco/__init__.py
@@ -1,0 +1,16 @@
+from ._tfmodisco import (
+    calculate_mean_expression_per_cell_type,
+    calculate_similarity_matrix,
+    create_pattern_matrix,
+    create_pattern_tf_dict,
+    create_tf_ct_matrix,
+    find_pattern,
+    find_pattern_matches,
+    generate_html_paths,
+    generate_nucleotide_sequences,
+    match_h5_files_to_classes,
+    pattern_similarity,
+    process_patterns,
+    read_motif_to_tf_file,
+    tfmodisco,
+)
